Closes-Bug: 1379513
Previously the glance-base script did not configure the registry
host IP. This causes API to try connecting to registry on
localhost. Since registry is bound to MY_IP, api is unable to
connect to registry. This change adds support for the glance-base
script to configure the registry_host parameter in api.conf to
MY_IP.

Change-Id: I1c711ff8aa95abec5ca306f03bd7e0a3883ca578