Adds support for Functional Nova-Network
Previsouly, the nova-network service was not fully functional.
This patch adds multi-interface support to the nova-network
container. The eth0 configuration is moved to a bridge (br100 per
nova defaults), and configures eth0 and eth1 as bridged
interfaces. eth0 is used for the nova-flat network and eth1 for
foating-ip's. This model is identical to typical bare-metal
nova-network deployments. The patch depends on the following patch
to kube-heat template:
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
!!!!!! PROVIDE PULL REQUEST!!!!!!!!